Work around Unicode encoding error on Windows
For Python <3.6 on Windows, encoding strings for output to the Windows
command prompt may result in a UnicodeEncodingError.  VOC does not take
encoding into consideration, and so the output differs.

Until such time as proper encoding support is implemented in VOC, use an
environment variable that Python provides for overriding the IO
encoding.  By setting this to UTF-8, the output may appear garbled, but
the error is avoided, and it matches the run-as-Java output.  For
consistency, pass the environment variable to Java as well.

Addresses #610 and #237.